adds listOfCategories js file refactores CategoryView
