+++ resources/listOfCategories.js
@@ -1,0 +1,35 @@
+const categories = [
+  {
+    id: 0,
+    category: 'Science',
+    description: 'Get to know latest news from the world of science',
+  },
+  {
+    id: 1,
+    category: 'Aliens',
+    description: 'Read undeniable proofs that we are not alone in the Universe',
+  },
+  {
+    id: 2,
+    category: 'Illuminati',
+    description: 'Get insides into a shadow government that pulls the strings',
+  },
+  {
+    id: 3,
+    category: 'Politics',
+    description:
+      'Get to know first what amendment is going to be violated next',
+  },
+  {
+    id: 4,
+    category: 'Covid',
+    description: 'All facts about the plot behind the pandemic',
+  },
+  {
+    id: 5,
+    category: 'Hollywood',
+    description: 'Truthful gossips about the life of celebrities',
+  },
+];
+
+export default categories;
