+++ views/CategoryView.jsx
@@ -1,40 +1,9 @@
 import React from 'react';
 import { StyleSheet, View, FlatList } from 'react-native';
 import CategoryListCard from '../components/CategoryListCard';
+import categories from '../resources/listOfCategories';
 
-const CategoryView = ({navigation}) => {
-  const categories = [
-    {
-      id: 0,
-      category: 'Science',
-      description: 'Get to know latest news from the world of science',
-    },
-    {
-      id: 1,
-      category: 'Aliens',
-      description: 'Read undeniable proofs that we are not alone in the Universe',
-    },
-    {
-      id: 2,
-      category: 'Illuminati',
-      description: 'Get insides into a shadow government that pulls the strings',
-    },
-    {
-      id: 3,
-      category: 'Politics',
-      description: 'Get to know first what amendment is going to be violated next',
-    },
-    {
-      id: 4,
-      category: 'Covid',
-      description: 'All facts about the plot behind the pandemic',
-    },
-    {
-      id: 5,
-      category: 'Hollywood',
-      description: 'Truthful gossips about the life of celebrities',
-    },
-  ];
+const CategoryView = ({ navigation }) => {
   return (
     <View testID='category-view' style={styles.container}>
       <FlatList
